# Avoids allocation in lazyFunc